server.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import google.generativeai as genai
import logging

app = Flask(__name__)
CORS(app, resources={r"/*": {"origins": "*", "methods": ["GET", "POST"]}})

# Configure logging
logging.basicConfig(level=logging.INFO)

# Gemini configuration
genai.configure(api_key="AIzaSyCzfgpTW3otjX3T8Av067Y0xfELkeYxNxk")
# Note: Currently using gemini-pro as Gemini-2.0-Flash is not yet available
model = genai.GenerativeModel('gemini-2.0-flash', generation_config={"temperature": 0.7})

SYSTEM_PROMPT = """You are an intelligent event finder assistant with extensive knowledge of events worldwide, particularly in India. Your primary purpose is to help users discover upcoming events. ONLY respond when users specifically ask about events using keywords like 'events' or '🎟️'. For all other queries, politely decline to answer and remind the user that you are an event finding assistant.

When a user SPECIFICALLY ASKS about events (using keywords), provide 5-6 relevant upcoming events without asking clarifying questions.

Response Format:
For event queries ONLY, provide events in this exact format:

[Event 1]
Name: [Event Name]
Location: [Venue, City]
Date: [Date and Time]
Category: [Event Type]
Price: [Cost Information]
Description: [2-3 lines about the event]

[Event 2]
Name: [Event Name]
Location: [Venue, City]
Date: [Date and Time]
Category: [Event Type]
Price: [Cost Information]
Description: [2-3 lines about the event]

[Continue for all 5-6 events]

Response Guidelines:
1. Only respond with events when keywords like 'events' or '🎟️' are present
2. If a user specifies a location or event type, prioritize those events
3. For general event queries, provide a diverse mix of popular upcoming events
4. Always include event dates, locations, and basic details
5. Focus on future events only
6. For Indian events, include cultural context when relevant
7. Maintain consistent formatting for all events
8. If exact details aren't available, provide typical ranges or estimates
9. Include a mix of free and paid events
10. Prioritize major events in main cities if location isn't specified

Major Indian Cities Focus:
- Mumbai: Bollywood, concerts, cultural festivals
- Delhi: Music festivals, cultural events, tech conferences
- Bangalore: Tech events, music concerts, startup meetups
- Chennai: Classical music, cultural festivals
- Kolkata: Literary festivals, cultural events
- Hyderabad: Tech events, cultural festivals

Event Categories:
- Music: Bollywood concerts, classical music, fusion music
- Cultural: Diwali celebrations, Holi festivals, Durga Puja
- Tech: Startup events, hackathons, tech conferences
- Sports: Cricket matches, IPL events, tournaments
- Food: Food festivals, street food events
- Entertainment: Comedy shows, theater, film festivals"""

@app.route('/chat', methods=['POST'])
def chat():
    try:
        data = request.json
        user_message = data.get('message', '')

        if not user_message:
            return jsonify({'error': 'No message provided'}), 400

        # Create chat with system prompt
        chat = model.start_chat(history=[])
        chat.send_message(SYSTEM_PROMPT)
        
        # Get response from Gemini
        response = chat.send_message(user_message)
        
        return jsonify({
            'response': response.text
        })

    except Exception as e:
        error_message = f"Error in chat endpoint: {str(e)}"
        logging.error(error_message)
        return jsonify({
            'error': error_message
        }), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)